src/Practico3/LinkedQueue/QueueTest.java
package Practico3.LinkedQueue;

import Practico3.EmptyQueueException;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class QueueTest {
    Queue<Integer> queue = new Queue<>();

    @Test
    void enqueue() throws EmptyQueueException {
        queue.enqueue(1);
        queue.enqueue(2);
        queue.enqueue(3);
        assertEquals(1, queue.dequeue().intValue());
    }

    @Test
    void dequeue() throws EmptyQueueException {
        queue.enqueue(1);
        queue.enqueue(2);
        queue.enqueue(3);
        assertEquals(1, queue.dequeue().intValue());
    }

    @Test
    void isEmpty() {
        assertTrue(queue.isEmpty());
    }

    @Test
    void isNotEmpty() {
        queue.enqueue(1);
        assertFalse(queue.isEmpty());
    }
}